Use thread::park_timeout
